Separated home and project controllers
